+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use AIDA64.SensorValues.Net to read shared memory
 - Changed display name to 'AIDA64'
 
+### Fixed
+
+- Handle TRIAL values
+
 ## 0.1.3 - 2023-07-27
 
 ### Changed